Remove legacy test runner infrastructure from build scripts and helix execution
