+++ src/tests/Common/CoreCLRTestLibrary/CoreCLRTestLibrary.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="AssertExtensions.cs" />
+    <Compile Include="CoreclrTestWrapperLib.cs" />
     <Compile Include="Generator.cs" />
     <Compile Include="Logging.cs" />
     <Compile Include="TestFramework.cs" />
@@ -19,8 +20,6 @@
     <Compile Include="Vectors.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Coreclr.TestWrapper\CoreclrTestWrapperLib.cs" Link="CoreclrTestWrapperLib.cs" />
-    <Compile Include="..\Coreclr.TestWrapper\MobileAppHandler.cs" Link="MobileAppHandler.cs" />
   </ItemGroup>
 
   <ItemGroup>
