+++ src/tests/Common/CoreCLRTestLibrary/OutOfProcessTest.cs
@@ -10,7 +10,6 @@ using System.Reflection;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using CoreclrTestLib;
 using Xunit;
 
 namespace TestLibrary
