--- src/tests/Common/Coreclr.TestWrapper/Coreclr.TestWrapper.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <NuGetTargetMoniker>$(NetCoreAppCurrentToolTargetFrameworkMoniker)</NuGetTargetMoniker>
-    <NuGetTargetMonikerShort>$(NetCoreAppToolCurrent)</NuGetTargetMonikerShort>
-    <Nullable>enable</Nullable>
-    <RunAnalyzers>true</RunAnalyzers>
-    <EnableAotAnalyzer>true</EnableAotAnalyzer>
-    <EnableTrimAnalyzer>true</EnableTrimAnalyzer>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="CoreclrTestWrapperLib.cs" />
-    <Compile Include="MobileAppHandler.cs" />
-  </ItemGroup>
-
-</Project>
