+++ src/tests/Common/XUnitLogChecker/XUnitLogChecker.csproj
@@ -15,9 +15,4 @@
     <Compile Include="$(MSBuildThisFileDirectory)XUnitLogChecker.cs" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)../Coreclr.TestWrapper/CoreclrTestWrapperLib.cs" Link="CoreclrTestWrapperLib.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)../Coreclr.TestWrapper/MobileAppHandler.cs" Link="MobileAppHandler.cs" />
-  </ItemGroup>
-
 </Project>
